Move authentication tests to a new test project.
