+++ .pipelines/perfview-job.yml
@@ -32,6 +32,7 @@ steps:
     testAssemblyVer2: |
       **\bin\**\LinuxTracingTests.dll
       **\bin\**\CtfTracingTests.dll
+      **\bin\**\SymbolAuthTests.dll
       **\bin\**\TraceEventTests.dll
       **\bin\**\PerfViewTests.dll
     testRunTitle: 'PerfView - ${{ parameters.flavor }}'
