+++ src/SymbolsAuth.Tests/app.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <system.diagnostics>
+    <trace>
+      <listeners>
+        <remove name="Default" />
+        <add name="ThrowingTraceListener" type="PerfView.TestUtilities.ThrowingTraceListener, PerfView.TestUtilities" />
+      </listeners>
+    </trace>
+  </system.diagnostics>
+</configuration>
