+++ src/SymbolsAuth.Tests/xunit.runner.json
@@ -1,0 +1,4 @@
+﻿{
+  "methodDisplay": "method",
+  "shadowCopy": false
+}
